fix: move setupPecas into config
